renamed userRouter to authRouter
